Move methods to helper and instance methods
